Forced type conversion for building blocks constructor materials, added tests.

+++ src/owimetadatabase_preprocessor/geometry/structures.py
@@ -152,7 +152,7 @@ class BuildingBlock(object):
         if "material" in json and self.subassembly:
             material_id = json["material"]
             for mat in self.subassembly.materials:
-                if mat.id == material_id:
+                if mat.id == np.int64(material_id):
                     self.material = mat
                     break
         self.json = json
